requirements.txt

# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-cloud-aiplatform>=1.30.0
evaluate>=0.4.0
prefect>=2.13.0
prefect-gcp>=0.4.0
gcsfs>=2023.6.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.0
datasets>=2.13.0
scikit-learn>=1.4.0
matplotlib>=3.7.0
wandb>=0.15.0
google-cloud-storage>=2.9.0
tensorflow>=2.12.0; platform_system!="Windows"
tensorflow-cpu>=2.12.0; platform_system=="Windows"
torch>=2.0.0
transformers>=4.30.0

# Redis dependencies
redis>=5.0.1

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
click>=8.1.0

# Testing and development
pytest==7.4.0
black==23.7.0
isort==5.12.0

# Optional integrations
flask
google-generativeai>=0.3.0
psutil
requests
selenium
pyyaml
